# Conflicts:
#	src/main/java/com/sparta/gimbap_heaven/GimbapHeavenApplication.java
#	src/main/java/com/sparta/gimbap_heaven/user/service/UserService.java

+++ src/main/java/com/sparta/gimbap_heaven/user/dto/UpdateMoneyRequestDto.java
@@ -6,5 +6,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class UpdateMoneyRequestDto {
-    private Integer money;
+    private double money;
 }
